requirements.txt
requests>=2.28
flake8>=6.0
radon>=5.0
bandit>=1.7
black==23.1.0
python-dotenv>=1.0
